Renaming j8 to jdk8
