+++ src/main/scala/com/fortysevendeg/scalacheck/datetime/instances/jdk8.scala
@@ -4,9 +4,9 @@ import com.fortysevendeg.scalacheck.datetime.typeclasses._
 import java.time._
 import java.time.temporal.ChronoUnit.MILLIS
 
-trait J8Instances {
+trait Jdk8Instances {
 
-  implicit val j8ForDuration: ScalaCheckDateTimeInfra[ZonedDateTime, Duration] = new ScalaCheckDateTimeInfra[ZonedDateTime, Duration] {
+  implicit val jdk8ForDuration: ScalaCheckDateTimeInfra[ZonedDateTime, Duration] = new ScalaCheckDateTimeInfra[ZonedDateTime, Duration] {
     def addRange(zonedDateTime: ZonedDateTime, duration: Duration): ZonedDateTime = zonedDateTime.plus(duration)
     def addMillis(zonedDateTime: ZonedDateTime, millis: Long): ZonedDateTime = zonedDateTime.plus(millis, MILLIS)
     def getMillis(zonedDateTime: ZonedDateTime): Long = zonedDateTime.toInstant.toEpochMilli
