+++ src/main/scala/com/fortysevendeg/scalacheck/datetime/instances/package.scala
@@ -2,5 +2,5 @@ package com.fortysevendeg.scalacheck.datetime
 
 package object instances {
   object joda extends JodaInstances
-  object j8 extends J8Instances
+  object jdk8 extends Jdk8Instances
 }
